server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
astroquery
astropy
